refactored appointment history jsx into sub components
